update the new conversation flow to fix the jumpy search
